% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drill_event_2023_mine_d}
\alias{drill_event_2023_mine_d}
\title{Log de Eventos 2023: Perfuratrizes Detalhado (Mina D)}
\format{
Tibble de eventos com colunas de GPS e Região.
}
\usage{
drill_event_2023_mine_d
}
\description{
Dados da aba "apontamentos 2023". Contém telemetria rica (GPS, Horímetro).
}
\seealso{
Other "Case Mina D": 
\code{\link{drill_event_log_mine_d}},
\code{\link{equipment_daily_act_mine_d}},
\code{\link{equipment_daily_plan_mine_d}},
\code{\link{production_daily_mine_d}}
}
\concept{"Case Mina D"}
\keyword{datasets}
