% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drill_event_mine_d}
\alias{drill_event_mine_d}
\title{Drilling Events from Mine D}
\format{
A tibble
}
\usage{
drill_event_mine_d
}
\description{
Drilling Events from Mine D
}
\keyword{datasets}
