% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drill_kpi_monthly_sn}
\alias{drill_kpi_monthly_sn}
\title{Indicadores Mensais: Perfuração (Serra Norte)}
\format{
Tibble mensal.
}
\usage{
drill_kpi_monthly_sn
}
\description{
Série histórica de KPIs (DF, UF, MTBF) para a frota de perfuratrizes.
}
\keyword{datasets}
