% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{equipment_daily_plan_mine_d}
\alias{equipment_daily_plan_mine_d}
\title{Planejamento Diário e Metas (Mina D)}
\format{
Tibble diária.
}
\usage{
equipment_daily_plan_mine_d
}
\description{
Dados da aba "programados". Budget de disponibilidade e produção.
}
\seealso{
Other "Case Mina D": 
\code{\link{drill_event_2023_mine_d}},
\code{\link{drill_event_log_mine_d}},
\code{\link{equipment_daily_act_mine_d}},
\code{\link{production_daily_mine_d}}
}
\concept{"Case Mina D"}
\keyword{datasets}
