% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_cycle_mine_a}
\alias{haul_cycle_mine_a}
\title{Ciclo de Transporte Detalhado (Mina A)}
\format{
Tibble de viagens:
\describe{
\item{id_ciclo}{Identificador sequencial.}
\item{TVV}{Tempo de Viagem Vazio (min).}
\item{TFC}{Tempo de Fila Carga (min).}
\item{TC}{Tempo de Carregamento (min).}
\item{massa_transportada}{Payload (ton).}
}
}
\usage{
haul_cycle_mine_a
}
\description{
Dados "Micro" (ciclo a ciclo) contendo a decomposição de tempos e produção.
Essencial para análise de tempos de fila e tempos fixos.
}
\seealso{
Other "Transporte e Ciclo": 
\code{\link{drill_event_mine_a}},
\code{\link{haul_daily_cep_mine_b}},
\code{\link{haul_daily_summary_mine_c}}
}
\concept{"Transporte e Ciclo"}
\keyword{datasets}
