% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_daily_cep_br}
\alias{haul_daily_cep_br}
\title{Histórico de Performance: Transporte (Brucutu)}
\format{
Tibble diária:
\describe{
\item{produtividade_ht}{Indicador de gestão: Produção / Horas Trabalhadas (t/h).}
\item{tempo_ciclo_total}{Soma de Tempos Fixos + Viagens (estimado via velocidade).}
\item{tempo_fixo}{Soma de filas, manobras e operações de carga/descarga.}
\item{HT}{Horas Trabalhadas (HEF + HAO + HTNP).}
\item{HEF}{Horas Efetivas (recalculado via UF e Atrasos).}
}
}
\usage{
haul_daily_cep_br
}
\description{
Dados diários de caminhões fora de estrada. Este dataset é rico pois reconstrói
o Tempo de Ciclo Total (TTC) e a Árvore de Horas (HT, HEF) a partir de variáveis
isoladas, permitindo exercícios de engenharia reversa.
}
\keyword{datasets}
