% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_daily_cep_mine_b}
\alias{haul_daily_cep_mine_b}
\title{Histórico de Performance: Transporte (Mina B)}
\format{
Tibble diária:
\describe{
\item{tkph}{Tonelada-Quilômetro por Hora.}
\item{tempo_ciclo_total}{TTC estimado.}
}
}
\usage{
haul_daily_cep_mine_b
}
\description{
Dados diários de caminhões com engenharia reversa de tempos.
}
\seealso{
Other "Transporte e Ciclo": 
\code{\link{drill_event_mine_a}},
\code{\link{haul_cycle_mine_a}},
\code{\link{haul_daily_summary_mine_c}}
}
\concept{"Transporte e Ciclo"}
\keyword{datasets}
