% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_daily_summary_mine_c}
\alias{haul_daily_summary_mine_c}
\title{Histórico Diário de Transporte (Mina C)}
\format{
Tibble diária.
}
\usage{
haul_daily_summary_mine_c
}
\description{
Dados consolidados diários da frota de transporte.
}
\seealso{
Other "Transporte e Ciclo": 
\code{\link{drill_event_mine_a}},
\code{\link{haul_cycle_mine_a}},
\code{\link{haul_daily_cep_mine_b}}
}
\concept{"Transporte e Ciclo"}
\keyword{datasets}
