% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_kpi_monthly_mine_a}
\alias{haul_kpi_monthly_mine_a}
\alias{drill_kpi_monthly_mine_a}
\alias{load_kpi_monthly_mine_a}
\alias{infra_kpi_monthly_mine_a}
\title{Indicadores Mensais Consolidados (Mina A)}
\format{
Tibble mensal.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1353 rows and 16 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3092 rows and 16 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4288 rows and 16 columns.
}
\usage{
haul_kpi_monthly_mine_a

drill_kpi_monthly_mine_a

load_kpi_monthly_mine_a

infra_kpi_monthly_mine_a
}
\description{
KPIs mensais oficiais (Requipam) para diversas frotas.
}
\seealso{
Other "Gestão e Estratégia": 
\code{\link{meta_dictionary}},
\code{\link{meta_event_classification}},
\code{\link{plan_budget_assumptions_mine_a}},
\code{\link{plan_daily_budget_mine_c}},
\code{\link{plan_daily_detailed_2021_mine_c}}

Other "Gestão e Estratégia": 
\code{\link{meta_dictionary}},
\code{\link{meta_event_classification}},
\code{\link{plan_budget_assumptions_mine_a}},
\code{\link{plan_daily_budget_mine_c}},
\code{\link{plan_daily_detailed_2021_mine_c}}

Other "Gestão e Estratégia": 
\code{\link{meta_dictionary}},
\code{\link{meta_event_classification}},
\code{\link{plan_budget_assumptions_mine_a}},
\code{\link{plan_daily_budget_mine_c}},
\code{\link{plan_daily_detailed_2021_mine_c}}

Other "Gestão e Estratégia": 
\code{\link{meta_dictionary}},
\code{\link{meta_event_classification}},
\code{\link{plan_budget_assumptions_mine_a}},
\code{\link{plan_daily_budget_mine_c}},
\code{\link{plan_daily_detailed_2021_mine_c}}
}
\concept{"Gestão e Estratégia"}
\keyword{datasets}
