% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_monthly_consolidated_mine_a}
\alias{haul_monthly_consolidated_mine_a}
\title{Monthly Haulage Consolidation for Mine A}
\format{
A tibble
}
\usage{
haul_monthly_consolidated_mine_a
}
\description{
Monthly Haulage Consolidation for Mine A
}
\keyword{datasets}
