% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_kpi_monthly_sn}
\alias{infra_kpi_monthly_sn}
\title{Indicadores Mensais: Infraestrutura (Serra Norte)}
\format{
Tibble mensal.
}
\usage{
infra_kpi_monthly_sn
}
\description{
Série histórica para Tratores de Esteira e Motoniveladoras (Patrol).
}
\keyword{datasets}
