% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_cost_plan}
\alias{infra_water_truck_cost_plan}
\title{Plano de Custos: Caminhão Pipa (Mina B)}
\format{
Tibble.
}
\usage{
infra_water_truck_cost_plan
}
\description{
Plano de Custos: Caminhão Pipa (Mina B)
}
\seealso{
Other "Infraestrutura e Apoio": 
\code{\link{drill_event_mine_a}},
\code{\link{infra_water_truck_events_mine_b}},
\code{\link{infra_water_truck_hourmeter_mine_b}},
\code{\link{infra_water_truck_task_list}},
\code{\link{maint_track_dozer_mine_a}}
}
\concept{"Infraestrutura e Apoio"}
\keyword{datasets}
