% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_events_br}
\alias{infra_water_truck_events_br}
\title{Histórico de Eventos: Caminhão Pipa (Brucutu)}
\format{
Tibble de eventos:
\describe{
\item{id_equipamento}{Identificador anonimizado (PIPA-XX).}
\item{duracao_h}{Duração do evento.}
\item{categoria}{HMC, MPS, etc.}
}
}
\usage{
infra_water_truck_events_br
}
\description{
Log de operações e manutenções da frota de irrigação (Infraestrutura).
Focado no estudo de otimização de periodicidade de preventiva.
}
\keyword{datasets}
