% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_hourmeter_br}
\alias{infra_water_truck_hourmeter_br}
\title{Horímetro Diário: Caminhão Pipa}
\format{
Tibble.
}
\usage{
infra_water_truck_hourmeter_br
}
\description{
Leitura de horímetro para cálculo de idade operacional.
}
\keyword{datasets}
