% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{load_daily_cep_mine_b}
\alias{load_daily_cep_mine_b}
\title{Histórico de Performance: Carga (Mina B)}
\format{
Tibble diária:
\describe{
\item{produtividade_ht}{Produtividade (t/HT).}
\item{HT}{Horas Trabalhadas Totais.}
\item{HEF}{Horas Efetivas.}
\item{HAO}{Horas de Atraso Operacional.}
}
}
\usage{
load_daily_cep_mine_b
}
\description{
Dados diários de escavadeiras seguindo a árvore de tempos padrão (GPV-M).
}
\seealso{
Other "Carga e Perfuração": 
\code{\link{drill_event_mine_a}}
}
\concept{"Carga e Perfuração"}
\keyword{datasets}
