% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{load_kpi_monthly_sn}
\alias{load_kpi_monthly_sn}
\title{Indicadores Mensais: Carga (Serra Norte)}
\format{
Tibble mensal.
}
\usage{
load_kpi_monthly_sn
}
\description{
Série histórica para Escavadeiras e Pás Carregadeiras.
}
\keyword{datasets}
