% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maint_track_dozer_mine_a}
\alias{maint_track_dozer_mine_a}
\title{Histórico de Manutenção: Tratores (Mina A)}
\format{
Tibble de eventos.
}
\usage{
maint_track_dozer_mine_a
}
\description{
Log detalhado de OS de tratores de esteira (Infraestrutura).
}
\seealso{
Other "Infraestrutura e Apoio": 
\code{\link{drill_event_mine_a}},
\code{\link{infra_water_truck_cost_plan}},
\code{\link{infra_water_truck_events_mine_b}},
\code{\link{infra_water_truck_hourmeter_mine_b}},
\code{\link{infra_water_truck_task_list}}
}
\concept{"Infraestrutura e Apoio"}
\keyword{datasets}
