% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_dictionary}
\alias{meta_dictionary}
\title{Dicionário de Termos (PT/EN/ES)}
\format{
Tibble.
}
\usage{
meta_dictionary
}
\description{
Glossário técnico trilíngue.
}
\seealso{
Other "Gestão e Estratégia": 
\code{\link{haul_kpi_monthly_mine_a}},
\code{\link{meta_event_classification}},
\code{\link{plan_budget_assumptions_mine_a}},
\code{\link{plan_daily_budget_mine_c}},
\code{\link{plan_daily_detailed_2021_mine_c}}
}
\concept{"Gestão e Estratégia"}
\keyword{datasets}
