% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_event_classification}
\alias{meta_event_classification}
\title{Tabela de Classificação de Eventos (Padrão PNR)}
\format{
Tibble.
}
\usage{
meta_event_classification
}
\description{
Regras de negócio para classificação de horas (HOI, HMC...).
}
\seealso{
Other "Gestão e Estratégia": 
\code{\link{haul_kpi_monthly_mine_a}},
\code{\link{meta_dictionary}},
\code{\link{plan_budget_assumptions_mine_a}},
\code{\link{plan_daily_budget_mine_c}},
\code{\link{plan_daily_detailed_2021_mine_c}}
}
\concept{"Gestão e Estratégia"}
\keyword{datasets}
