% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plan_budget_assumptions_sn}
\alias{plan_budget_assumptions_sn}
\title{Premissas Orçamentárias Anuais (Serra Norte)}
\format{
Tibble mensal:
\describe{
\item{data}{Mês de referência.}
\item{id_frota}{Identificador da frota (ex: FROTA-A).}
\item{num_caminhoes_plan}{Quantidade de ativos orçada.}
\item{dmt_plan}{Distância Média de Transporte planejada.}
\item{DF_plan}{Meta de Disponibilidade Física.}
}
}
\source{
Planilhas de Premissas Orçamentárias (Engenharia de Minas).
}
\usage{
plan_budget_assumptions_sn
}
\description{
Dados consolidados do orçamento (Budget) anual, detalhados mês a mês por frota.
Contém as metas físicas (DMT, Produção) e de disponibilidade que nortearam o ano.
}
\keyword{datasets}
