% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plan_daily_budget_it}
\alias{plan_daily_budget_it}
\title{Metas Diárias de Transporte (Itabira)}
\format{
Tibble diária:
\describe{
\item{data}{Data da meta.}
\item{DF_plan}{Meta de Disponibilidade (\%).}
\item{producao_plan}{Meta de Produção (ton).}
\item{num_caminhoes_plan}{Dimensionamento de frota previsto.}
}
}
\usage{
plan_daily_budget_it
}
\description{
Orçamento (Budget) diário para a frota de transporte.
Deve ser cruzado com \code{haul_daily_summary_it} para análise de aderência (Real x Meta).
}
\keyword{datasets}
