% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_event_log_mine_e}
\alias{plant_event_log_mine_e}
\title{Log de Eventos: Pátio e Usina (Mina E)}
\format{
Tibble com histórico de eventos:
\describe{
\item{tipo_ativo}{Tipo de máquina (Empilhadeira, Recuperadora, Correia).}
\item{id_equipamento}{Tag anonimizada (STK=Stacker, RCL=Reclaimer).}
\item{sistema}{Sistema produtivo (ex: Circuito Pen. Primário).}
\item{categoria_kpi}{Classificação de horas (HOI, HMC, HEF).}
\item{causa_basica}{Causa primária da parada.}
}
}
\source{
Dados anonimizados de operação em Serra Sul (S11D).
}
\usage{
plant_event_log_mine_e
}
\description{
Dataset de grande volumetria cobrindo o Beneficiamento e Pátio de Regularização da Mina E.
Contém dados de Empilhadeiras, Recuperadoras e circuitos de peneiramento.
}
\seealso{
Other "Case Mina E": 
\code{\link{ipcc_event_log_mine_e}}

Other "Beneficiamento e Processamento": 
\code{\link{ipcc_event_log_mine_e}}
}
\concept{"Beneficiamento e Processamento"}
\concept{"Case Mina E"}
\keyword{datasets}
