% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plant_production_mine_f}
\alias{plant_production_mine_f}
\title{Produção Horária e Taxas de Alimentação (Mina F)}
\format{
Tibble com 41.769 registos:
\describe{
\item{inicio, fim}{Janelas temporais, tipicamente de 1 hora.}
\item{id_equipamento}{Tag do equipamento de alimentação/britagem.}
\item{tonelagem}{Quantidade de minério processada no intervalo (ton).}
\item{origem, destino}{Pontos de transferência no fluxo da usina.}
}
}
\source{
L:/Meu Drive/Dados_Confidenciais/data_salobo_taxa.Rdata
}
\usage{
plant_production_mine_f
}
\description{
Registos horários de tonelagem e taxa (t/h). Cobre a Britagem Primária,
o Circuito Singelo e as 4 Linhas de Peneiramento/Moagem. Essencial para
cálculos de performance e OEE.
}
\seealso{
Other "Case Mina F": 
\code{\link{inventory_raw_mine_f}},
\code{\link{plant_detailed_mine_f}}
}
\concept{"Case Mina F"}
\concept{"Performance"}
\keyword{datasets}
