% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production_daily_mine_d}
\alias{production_daily_mine_d}
\title{Produção Diária e Qualidade (Mina D)}
\format{
Tibble diária.
}
\usage{
production_daily_mine_d
}
\description{
Dados da aba "D+30 infra". Balanço de massa, teor de Cobre e REM.
}
\seealso{
Other "Case Mina D": 
\code{\link{drill_event_2023_mine_d}},
\code{\link{drill_event_log_mine_d}},
\code{\link{equipment_daily_act_mine_d}},
\code{\link{equipment_daily_plan_mine_d}}
}
\concept{"Case Mina D"}
\keyword{datasets}
