% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_daily_summary_mine_c}
\alias{haul_daily_summary_mine_c}
\title{Hist칩rico Di치rio de Transporte (Mina C)}
\format{
Tibble di치ria.
}
\usage{
haul_daily_summary_mine_c
}
\description{
Dados consolidados di치rios da frota de transporte.
}
\keyword{datasets}
