% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{haul_maint_costs_mine_a}
\alias{haul_maint_costs_mine_a}
\title{Histórico de Custos Detalhados (Mina A)}
\format{
Tibble mensal:
\describe{
\item{data}{Mês de competência.}
\item{custo_total}{Soma dos custos do equipamento.}
\item{custo_pecas}{Valor gasto em materiais.}
}
}
\usage{
haul_maint_costs_mine_a
}
\description{
Dados financeiros mensais por equipamento individual.
}
\keyword{datasets}
