% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_cost_plan}
\alias{infra_water_truck_cost_plan}
\title{Plano de Custos: Caminhão Pipa (Mina B)}
\format{
Tibble.
}
\usage{
infra_water_truck_cost_plan
}
\description{
Plano de Custos: Caminhão Pipa (Mina B)
}
\keyword{datasets}
