% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_cost_plan}
\alias{infra_water_truck_cost_plan}
\title{Plano de Custos de Preventiva (Caminhão Pipa)}
\format{
Tibble.
}
\usage{
infra_water_truck_cost_plan
}
\description{
Lista de peças e custos associados a cada tipo de revisão (300h, 600h...).
Permite calcular o custo total de cada estratégia de manutenção.
}
\keyword{datasets}
