% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_events_mine_b}
\alias{infra_water_truck_events_mine_b}
\title{Histórico de Eventos: Caminhão Pipa (Mina B)}
\format{
Tibble de eventos.
}
\usage{
infra_water_truck_events_mine_b
}
\description{
Log de operações da frota de irrigação.
}
\keyword{datasets}
