% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_hourmeter_mine_b}
\alias{infra_water_truck_hourmeter_mine_b}
\title{Horímetro: Caminhão Pipa (Mina B)}
\format{
Tibble.
}
\usage{
infra_water_truck_hourmeter_mine_b
}
\description{
Horímetro: Caminhão Pipa (Mina B)
}
\keyword{datasets}
