% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infra_water_truck_task_list}
\alias{infra_water_truck_task_list}
\title{Lista de Tarefas: Caminhão Pipa (Mina B)}
\format{
Tibble.
}
\usage{
infra_water_truck_task_list
}
\description{
Lista de Tarefas: Caminhão Pipa (Mina B)
}
\keyword{datasets}
