% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipcc_event_log_mine_e}
\alias{ipcc_event_log_mine_e}
\title{Log de Eventos: IPCC / Truckless (Mina E)}
\format{
Tibble com histórico de eventos:
\describe{
\item{metodo_lavra}{Identifica o método (IPCC Truckless).}
\item{tipo_ativo}{Classificação do equipamento (Correia, Britador Móvel, Alimentador).}
\item{id_equipamento}{Tag anonimizada (CNV=Conveyor, CRH=Crusher).}
\item{sistema}{Parte do circuito (Estéril ou Minério).}
\item{categoria_kpi}{Classificação de horas (HOI, HMC, HEF).}
\item{impacta_indicador}{Se o evento parou a produção (Loss).}
}
}
\source{
Dados anonimizados de operação em Serra Sul (S11D).
}
\usage{
ipcc_event_log_mine_e
}
\description{
Dataset de alta frequência de uma mina operada via IPCC (In-Pit Crushing and Conveying).
}
\details{
Diferente das minas tradicionais (Caminhão/Escavadeira), este sistema é contínuo.
As falhas em um equipamento (ex: Britador Móvel) propagam-se imediatamente para as correias (Sistemas em Série).
}
\seealso{
Other "Case Mina E": 
\code{\link{plant_event_log_mine_e}}

Other "Beneficiamento e Processamento": 
\code{\link{plant_event_log_mine_e}}
}
\concept{"Beneficiamento e Processamento"}
\concept{"Case Mina E"}
\keyword{datasets}
