% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maint_orders_mine_b}
\alias{maint_orders_mine_b}
\title{Ordens de Serviço: Manutenção (Mina B)}
\format{
Tibble de eventos.
}
\usage{
maint_orders_mine_b
}
\description{
Histórico textual de falhas (2021-2022).
}
\keyword{datasets}
