% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maint_stops_mine_b}
\alias{maint_stops_mine_b}
\title{Log Completo de Paradas (Mina B)}
\format{
Tibble de eventos.
}
\usage{
maint_stops_mine_b
}
\description{
Dataset que diferencia Manutenção (HMC) de Ociosidade (HOI).
}
\keyword{datasets}
