% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maint_track_dozer_mine_a}
\alias{maint_track_dozer_mine_a}
\title{Histórico de Manutenção: Tratores (Mina A)}
\format{
Tibble de eventos.
}
\usage{
maint_track_dozer_mine_a
}
\description{
Log detalhado de OS de tratores de esteira (Infraestrutura).
}
\keyword{datasets}
