% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maint_truck_events_mine_c}
\alias{maint_truck_events_mine_c}
\title{Histórico de Manutenção Preventiva (Mina C)
Case de Otimização de Periodicidade.}
\format{
Tibble de eventos.
}
\usage{
maint_truck_events_mine_c
}
\description{
Histórico de Manutenção Preventiva (Mina C)
Case de Otimização de Periodicidade.
}
\keyword{datasets}
