% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maint_truck_pautas_mine_c}
\alias{maint_truck_pautas_mine_c}
\title{Histórico de Pautas (Mina C)}
\format{
Tibble.
}
\usage{
maint_truck_pautas_mine_c
}
\description{
Histórico de Pautas (Mina C)
}
\keyword{datasets}
