% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_event_classification}
\alias{meta_event_classification}
\title{Tabela de Classificação de Eventos (Padrão PNR)}
\format{
Tibble.
}
\usage{
meta_event_classification
}
\description{
Regras de negócio para classificação de horas (HOI, HMC...).
}
\keyword{datasets}
