% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plan_budget_assumptions_mine_a}
\alias{plan_budget_assumptions_mine_a}
\title{Premissas Orçamentárias Anuais (Mina A)}
\format{
Tibble mensal:
\describe{
\item{DF_plan}{Meta de Disponibilidade Física.}
\item{producao_plan}{Meta de produção.}
}
}
\usage{
plan_budget_assumptions_mine_a
}
\description{
Dados consolidados do orçamento (Budget) anual.
}
\keyword{datasets}
