% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plan_daily_budget_mine_c}
\alias{plan_daily_budget_mine_c}
\title{Metas Diárias de Transporte (Mina C)}
\format{
Tibble diária.
}
\usage{
plan_daily_budget_mine_c
}
\description{
Orçamento (Budget) diário para aderência.
}
\keyword{datasets}
