% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plan_daily_detailed_2021_mine_c}
\alias{plan_daily_detailed_2021_mine_c}
\title{Metas Detalhadas de Transporte 2021 (Mina C)}
\format{
Tibble diária.
}
\usage{
plan_daily_detailed_2021_mine_c
}
\description{
Budget com granularidade de tempos de ciclo.
}
\keyword{datasets}
