% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production_forecast_mine_d}
\alias{production_forecast_mine_d}
\title{Produção Diária e Balanço de Massa (Mina D)}
\format{
Um data frame (tibble) com as seguintes colunas:
\describe{
\item{date}{Data da operação.}
\item{total_movement_t}{Movimentação total de massa em toneladas (Minério + Estéril + Remanejo).}
\item{ore_mined_t}{Produção de Run of Mine (ROM) em toneladas (TBN - Tonelada Bruta Húmida).}
\item{waste_mined_t}{Massa de estéril lavrada no dia.}
\item{crusher_feed_t}{Massa total alimentada no britador primário.}
\item{rehandling_t}{Massa movimentada em pilhas de estoque (Remanejo).}
\item{grade_cu_percent}{Teor de Cobre (Cu) realizado no ROM em percentagem.}
\item{grade_au_gpt}{Teor de Ouro (Au) realizado em gramas por tonelada (g/t).}
\item{contained_cu_t}{Cobre contido em toneladas.}
\item{contained_au_oz}{Ouro contido (frequentemente expresso em onças ou massa TBS).}
}
}
\source{
Gestão de Performance da Unidade (Dados Reais Tratados)
}
\usage{
production_forecast_mine_d
}
\description{
Registos diários históricos da produção de minério, remoção de estéril e
alimentação da planta. Essencial para análises de aderência ao plano e
balanço de massas no contexto de Mining Analytics.
}
\keyword{datasets}
